pipeline/requirements.txt
pandas
PyMySQL
python-dotenv
awswrangler